+++ src/main/java/nextstep/security/core/uesrdetails/UserDetails.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.core.uesrdetails;
 
 public interface UserDetails {
     String getUsername();
