+++ src/main/java/nextstep/security/core/uesrdetails/UserDetailsService.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.core.uesrdetails;
 
 public interface UserDetailsService {
     UserDetails loadUserByUsername(String username);
